Create ultrasonic.ino
Create a system with the proximity/distance sensor. You should use the sensor to get the
information about the distance to an obstacle. You should use four LEDs to indicate how close the
obstacle is from the sensor.
If the object is out of range, then no LEDs should be turn on. If the object is in range, you should turn
on 1, 2 or 3 LEDs, depending on the distance. If the object is too close, you should turn on four LEDs.