Create coolBitCounter.c
Write a program that counts the number of 0s and 1s in the binary representation of a decimal
number.
Your program should ask the user to provide a decimal number. Then it should print the decimal
number, it's binary representation and then count the number of 0s and 1s.
The interaction with the program should look like this: