Create advancedRemoveShortest.c
Write a program in C that reads 10 strings from a console, stores them in an array and finds the
shortest string in the array.

The program should contain a function that takes as parameter the array, find the shortest string,
remove it and return the array that is shorter by one element.

The program should read the strings from the console.

The program should have a separate function for finding and removing the string.

The program should write the string that has been removed to the console.

The program should
also write all elements of the array before and after removal of the shortest string.